backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
fastapi==0.113.0
h11==0.14.0
idna==3.8
pydantic==2.9.0
pydantic_core==2.23.2
sniffio==1.3.1
starlette==0.38.4
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
mysql-connector-python== 9.0.0
bcrypt==4.2.0
openai==1.45.0
python-dotenv==1.0.1
sqlalchemy==2.0.34
pymysql
passlib[bcrypt]
python-jose
email-validator
authlib==1.3.2 
python-dotenv
starlette
itsdangerous==2.2.0
firebase-adminopenai
pinecone
langchain
langchain-community
pypdf
azure-ai-inference
